pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "access_py_telemetry"
version = "0.1.0"
description = "Contains ipython extensions to automatically add telemetry to registered function calls from Ipython kernels"
readme = "README.md"
authors = [
  { name = "ACCESS-NRI" },
]
maintainers = [
  { name = "ACCESS-NRI" },
]
classifiers = [

]
license = {text = "Apache Software License 2.0"}
dependencies = [
   "ipython>=7.0.0",
   "pydantic>=2.0.0",
   "pyyaml>=6.0.0",
   "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "coverage",  # testing
    "mypy",  # linting
    "pytest",  # testing
    "ruff"  # linting
]

[project.scripts]
access-py-telemetry = "access_py_telemetry.cli:configure_telemetry"

[project.urls]

bugs = "https://github.com/ACCESS-NRI/access-py-telemetry/issues"
changelog = "https://github.com/ACCESS-NRI/access-py-telemetry/blob/main/CHANGELOG.md"
homepage = "https://github.com/ACCESS-NRI/access-py-telemetry"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["*.*"]




# Mypy
# ----

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

